// src/model/AnalyticsPlays400Error.ts
/**
 * @api.video/nodejs-client
 * api.video is an API that encodes on the go to facilitate immediate playback, enhancing viewer streaming experiences across multiple devices and platforms. You can stream live or on-demand online videos within minutes.
 *
 * The version of the OpenAPI document: 1
 *
 *
 * NOTE: This class is auto generated.
 * Do not edit the class manually.
 */

import AttributeType from './AttributeType.js';

export default class AnalyticsPlays400Error {
  /**
   * A link to the error documentation.
   */
  'type'?: string;
  /**
   * A description of the error that occurred.
   */
  'title'?: string;
  /**
   * The HTTP status code.
   */
  'status'?: number;
  /**
   * A solution for the error.
   */
  'detail'?: string;
  /**
   * The name of the parameter that caused the error.
   */
  'name'?: string;

  static readonly discriminator?: string = undefined;

  static readonly attributeTypeMap: Array<AttributeType> = [
    {
      name: 'type',
      baseName: 'type',
      type: 'string',
      format: '',
    },
    {
      name: 'title',
      baseName: 'title',
      type: 'string',
      format: '',
    },
    {
      name: 'status',
      baseName: 'status',
      type: 'number',
      format: '',
    },
    {
      name: 'detail',
      baseName: 'detail',
      type: 'string',
      format: '',
    },
    {
      name: 'name',
      baseName: 'name',
      type: 'string',
      format: '',
    },
  ];

  static getAttributeTypeMap(): Array<AttributeType> {
    return AnalyticsPlays400Error.attributeTypeMap;
  }
}