// src/model/BadRequest.ts
/**
 * @api.video/nodejs-client
 * api.video is an API that encodes on the go to facilitate immediate playback, enhancing viewer streaming experiences across multiple devices and platforms. You can stream live or on-demand online videos within minutes.
 *
 * The version of the OpenAPI document: 1
 *
 *
 * NOTE: This class is auto generated.
 * Do not edit the class manually.
 */

import AttributeType from './AttributeType';

export default class BadRequest {
  'type'?: string;
  'title'?: string;
  'name'?: string;
  'status'?: number;
  'problems'?: Array<BadRequest>;

  static readonly discriminator?: string = undefined;

  static readonly attributeTypeMap: Array<AttributeType> = [
    {
      name: 'type',
      baseName: 'type',
      type: 'string',
      format: '',
    },
    {
      name: 'title',
      baseName: 'title',
      type: 'string',
      format: '',
    },
    {
      name: 'name',
      baseName: 'name',
      type: 'string',
      format: '',
    },
    {
      name: 'status',
      baseName: 'status',
      type: 'number',
      format: '',
    },
    {
      name: 'problems',
      baseName: 'problems',
      type: 'Array<BadRequest>',
      format: '',
    },
  ];

  static getAttributeTypeMap(): Array<AttributeType> {
    return BadRequest.attributeTypeMap;
  }
}