// src/model/PaginationLink.ts
/**
 * @api.video/nodejs-client
 * api.video is an API that encodes on the go to facilitate immediate playback, enhancing viewer streaming experiences across multiple devices and platforms. You can stream live or on-demand online videos within minutes.
 *
 * The version of the OpenAPI document: 1
 *
 *
 * NOTE: This class is auto generated.
 * Do not edit the class manually.
 */

import AttributeType from './AttributeType.js';

export default class PaginationLink {
  'rel'?: string;
  'uri'?: string;

  static readonly discriminator?: string = undefined;

  static readonly attributeTypeMap: Array<AttributeType> = [
    {
      name: 'rel',
      baseName: 'rel',
      type: 'string',
      format: '',
    },
    {
      name: 'uri',
      baseName: 'uri',
      type: 'string',
      format: 'uri',
    },
  ];

  static getAttributeTypeMap(): Array<AttributeType> {
    return PaginationLink.attributeTypeMap;
  }
}